backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
transformers
torch
python-dotenv
sentencepiece

